+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-08-19
+
+### Removed
+
+- Metabase visualisation support and code.
+
 ## 2020-08-18
 
 - IE11 fixes for Your files page